Deleted the redundant flightresults container
